+++ src/components/create-safe/logic/index.ts
@@ -11,12 +11,11 @@ import {
 } from '@/services/contracts/safeContracts'
 import { LATEST_SAFE_VERSION } from '@/config/constants'
 import type { PredictSafeProps } from '@gnosis.pm/safe-core-sdk/dist/src/safeFactory'
-import type { SafeFormData } from '@/components/create-safe/types'
+import type { SafeFormData, PendingSafeTx } from '@/components/create-safe/types.d'
 import type { ConnectedWallet } from '@/services/onboard'
 import { BigNumber } from '@ethersproject/bignumber'
 import { getSafeInfo } from '@gnosis.pm/safe-react-gateway-sdk'
 import { backOff } from 'exponential-backoff'
-import type { PendingSafeTx } from '@/components/create-safe'
 import { SafeCreationStatus } from '@/components/create-safe/status/useSafeCreation'
 import { didRevert, type EthersError } from '@/utils/ethers-utils'
 import { Errors, logError } from '@/services/exceptions'
