+++ src/components/create-safe/status/__tests__/useSafeCreationEffects.test.ts
@@ -7,7 +7,7 @@ import * as web3 from '@/hooks/wallets/web3'
 import * as pendingSafe from '@/components/create-safe/logic'
 import * as chainIdModule from '@/hooks/useChainId'
 import { Web3Provider } from '@ethersproject/providers'
-import type { PendingSafeData } from '@/components/create-safe'
+import type { PendingSafeData } from '@/components/create-safe/types.d'
 import { AppRoutes } from '@/config/routes'
 import { CONFIG_SERVICE_CHAINS } from '@/tests/mocks/chains'
 import useSafeCreationEffects from '@/components/create-safe/status/useSafeCreationEffects'
