+++ src/components/create-safe/useSetCreationStep.ts
@@ -3,10 +3,10 @@ import useIsWrongChain from '@/hooks/useIsWrongChain'
 import useWallet from '@/hooks/wallets/useWallet'
 import { useEffect } from 'react'
 import useLocalStorage from '@/services/local-storage/useLocalStorage'
-import type { PendingSafeData } from '@/components/create-safe/index'
 import { SAFE_PENDING_CREATION_STORAGE_KEY } from '@/components/create-safe/status/CreationStatus'
+import type { PendingSafeData } from '@/components/create-safe/types.d'
 
-const useResetSafeCreation = (setStep: StepRenderProps['setStep']) => {
+const useSetCreationStep = (setStep: StepRenderProps['setStep']) => {
   const [pendingSafe] = useLocalStorage<PendingSafeData | undefined>(SAFE_PENDING_CREATION_STORAGE_KEY, undefined)
   const wallet = useWallet()
   const isWrongChain = useIsWrongChain()
@@ -23,4 +23,4 @@ const useResetSafeCreation = (setStep: StepRenderProps['setStep']) => {
   }, [setStep, isWrongChain, wallet, pendingSafe])
 }
 
-export default useResetSafeCreation
+export default useSetCreationStep
