+++ src/components/tx/TxStepper/vertical.tsx
@@ -7,13 +7,12 @@ import { StepContent } from '@mui/material'
 import type { TxStepperProps } from '@/components/tx/TxStepper/useTxStepper'
 import { useTxStepper } from '@/components/tx/TxStepper/useTxStepper'
 
-const VerticalTxStepper = ({ steps, initialData, initialStep, onClose, onFinish }: TxStepperProps): ReactElement => {
+const VerticalTxStepper = ({ steps, initialData, initialStep, onClose }: TxStepperProps): ReactElement => {
   const { onBack, onSubmit, setStep, activeStep, stepData } = useTxStepper({
     steps,
     initialData,
     initialStep,
     onClose,
-    onFinish,
   })
 
   return (
